Rename shared utilities directory to 'common'
